requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
setuptools
gunicorn
whitenoise
dj-database-url